requirements.txt
pandas
scikit-learn
numpy
pandas_gbq
matplotlib
google-cloud-storage
google-cloud-bigquery-storage
flask
pyarrow